refactor app.py (3/7): extract git helpers to helpers/git.py
get_git_status and perform_git_ops move to helpers/git.py. No
constants needed — both shell out via subprocess and rely on cwd
being the repo root (unchanged behavior).

Smoke-tested: /publish renders 200 (calls get_git_status for the
HUD shield indicator). perform_git_ops not exercised locally — it
would push, only runs on PA via /publish POST.
